Reorganize ``api_fastapi`` folder into apps
In preparation for adding https://github.com/apache/airflow/pull/43015, this PRs re-organizes the ``api_fastapi``. The goal will be to have
2 apps: `execution_api` (Task Execution API: AIP-72) and the `core_api` (UI and Public APi).
