+++ airflow/api_fastapi/app.py
@@ -17,40 +17,25 @@
 from __future__ import annotations
 
 import logging
-import os
-from pathlib import Path
-from typing import cast
 
-from fastapi import FastAPI, Request
+from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
-from fastapi.responses import HTMLResponse
-from fastapi.staticfiles import StaticFiles
-from fastapi.templating import Jinja2Templates
 
-from airflow.settings import AIRFLOW_PATH
-from airflow.www.extensions.init_dagbag import get_dag_bag
+from airflow.api_fastapi.core_api.app import init_dag_bag, init_plugins, init_views
 
 log = logging.getLogger(__name__)
 
 app: FastAPI | None = None
 
 
-def init_dag_bag(app: FastAPI) -> None:
-    """
-    Create global DagBag for the FastAPI application.
-
-    To access it use ``request.app.state.dag_bag``.
-    """
-    app.state.dag_bag = get_dag_bag()
-
-
-def create_app() -> FastAPI:
+def create_app(apps: list[str] | None = None) -> FastAPI:
     from airflow.configuration import conf
 
     app = FastAPI(
+        title="Airflow API",
         description="Airflow API. All endpoints located under ``/public`` can be used safely, are stable and backward compatible. "
         "Endpoints located under ``/ui`` are dedicated to the UI and are subject to breaking change "
-        "depending on the need of the frontend. Users should not rely on those but use the public ones instead."
+        "depending on the need of the frontend. Users should not rely on those but use the public ones instead.",
     )
 
     init_dag_bag(app)
@@ -73,60 +58,6 @@ def create_app() -> FastAPI:
         )
 
     return app
-
-
-def init_views(app: FastAPI) -> None:
-    """Init views by registering the different routers."""
-    from airflow.api_fastapi.routes.public import public_router
-    from airflow.api_fastapi.routes.ui import ui_router
-
-    app.include_router(ui_router)
-    app.include_router(public_router)
-
-    dev_mode = os.environ.get("DEV_MODE", False) == "true"
-
-    directory = Path(AIRFLOW_PATH) / ("airflow/ui/dev" if dev_mode else "airflow/ui/dist")
-
-    # During python tests or when the backend is run without having the frontend build
-    # those directories might not exist. App should not fail initializing in those scenarios.
-    Path(directory).mkdir(exist_ok=True)
-
-    templates = Jinja2Templates(directory=directory)
-
-    app.mount(
-        "/static",
-        StaticFiles(
-            directory=directory,
-            html=True,
-        ),
-        name="webapp_static_folder",
-    )
-
-    @app.get("/webapp/{rest_of_path:path}", response_class=HTMLResponse, include_in_schema=False)
-    def webapp(request: Request, rest_of_path: str):
-        return templates.TemplateResponse("/index.html", {"request": request}, media_type="text/html")
-
-
-def init_plugins(app: FastAPI) -> None:
-    """Integrate FastAPI app plugins."""
-    from airflow import plugins_manager
-
-    plugins_manager.initialize_fastapi_plugins()
-
-    # After calling initialize_fastapi_plugins, fastapi_apps cannot be None anymore.
-    for subapp_dict in cast(list, plugins_manager.fastapi_apps):
-        name = subapp_dict.get("name")
-        subapp = subapp_dict.get("app")
-        if subapp is None:
-            log.error("'app' key is missing for the fastapi app: %s", name)
-            continue
-        url_prefix = subapp_dict.get("url_prefix")
-        if url_prefix is None:
-            log.error("'url_prefix' key is missing for the fastapi app: %s", name)
-            continue
-
-        log.debug("Adding subapplication %s under prefix %s", name, url_prefix)
-        app.mount(url_prefix, subapp)
 
 
 def cached_app(config=None, testing=False) -> FastAPI:
