+++ airflow/api_fastapi/core_api/openapi/__init__.py
