+++ airflow/api_fastapi/core_api/routes/public/connections.py
@@ -21,11 +21,14 @@ from sqlalchemy import select
 from sqlalchemy.orm import Session
 from typing_extensions import Annotated
 
-from airflow.api_fastapi.db.common import get_session, paginated_select
-from airflow.api_fastapi.openapi.exceptions import create_openapi_http_exception_doc
-from airflow.api_fastapi.parameters import QueryLimit, QueryOffset, SortParam
-from airflow.api_fastapi.routes.router import AirflowRouter
-from airflow.api_fastapi.serializers.connections import ConnectionCollectionResponse, ConnectionResponse
+from airflow.api_fastapi.common.db.common import get_session, paginated_select
+from airflow.api_fastapi.common.parameters import QueryLimit, QueryOffset, SortParam
+from airflow.api_fastapi.common.router import AirflowRouter
+from airflow.api_fastapi.core_api.openapi.exceptions import create_openapi_http_exception_doc
+from airflow.api_fastapi.core_api.serializers.connections import (
+    ConnectionCollectionResponse,
+    ConnectionResponse,
+)
 from airflow.models import Connection
 
 connections_router = AirflowRouter(tags=["Connection"], prefix="/connections")
