+++ airflow/api_fastapi/core_api/serializers/variables.py
