+++ tests/api_fastapi/core_api/routes/public/__init__.py
@@ -14,13 +14,3 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-from __future__ import annotations
-
-from airflow.api_fastapi.routes.router import AirflowRouter
-from airflow.api_fastapi.routes.ui.assets import assets_router
-from airflow.api_fastapi.routes.ui.dashboard import dashboard_router
-
-ui_router = AirflowRouter(prefix="/ui")
-
-ui_router.include_router(assets_router)
-ui_router.include_router(dashboard_router)
