+++ tests/api_fastapi/core_api/routes/public/test_dags.py
@@ -293,7 +293,7 @@ class TestDagDetails(TestDagEndpoint):
             "description": None,
             "doc_md": "details",
             "end_date": None,
-            "fileloc": "/opt/airflow/tests/api_fastapi/routes/public/test_dags.py",
+            "fileloc": "/opt/airflow/tests/api_fastapi/core_api/routes/public/test_dags.py",
             "file_token": file_token,
             "has_import_errors": False,
             "has_task_concurrency_limits": True,
@@ -358,7 +358,7 @@ class TestGetDag(TestDagEndpoint):
             "dag_id": dag_id,
             "dag_display_name": dag_display_name,
             "description": None,
-            "fileloc": "/opt/airflow/tests/api_fastapi/routes/public/test_dags.py",
+            "fileloc": "/opt/airflow/tests/api_fastapi/core_api/routes/public/test_dags.py",
             "file_token": file_token,
             "is_paused": False,
             "is_active": True,
