+++ tests/api_fastapi/core_api/routes/public/test_monitor.py
