+++ tests/api_fastapi/core_api/routes/ui/__init__.py
@@ -14,17 +14,3 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-
-from __future__ import annotations
-
-from airflow.api.common.airflow_health import get_airflow_health
-from airflow.api_fastapi.routes.router import AirflowRouter
-from airflow.api_fastapi.serializers.monitor import HealthInfoSchema
-
-monitor_router = AirflowRouter(tags=["Monitor"], prefix="/monitor")
-
-
-@monitor_router.get("/health")
-async def get_health() -> HealthInfoSchema:
-    airflow_health_status = get_airflow_health()
-    return HealthInfoSchema.model_validate(airflow_health_status)
