+++ tests/api_fastapi/core_api/routes/ui/test_assets.py
